addresses #89 by moving all client stuff to the client crate
