+++ icann-rdap-common/Cargo.toml
@@ -17,7 +17,6 @@ idna.workspace = true
 ipnet.workspace = true
 lazy_static.workspace = true
 prefix-trie.workspace = true
-reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 strum.workspace = true
