+++ icann-rdap-srv/src/error.rs
@@ -6,11 +6,8 @@ use axum::{
 };
 use envmnt::errors::EnvmntError;
 use http::StatusCode;
-use icann_rdap_client::RdapClientError;
-use icann_rdap_common::{
-    iana::IanaResponseError,
-    response::{types::Common, RdapResponse, RdapResponseError},
-};
+use icann_rdap_client::{iana_request::IanaResponseError, RdapClientError};
+use icann_rdap_common::response::{types::Common, RdapResponse, RdapResponseError};
 use ipnet::PrefixLenError;
 use thiserror::Error;
